+++ novokhatskiy.danil/S5/main.cpp
@@ -3,8 +3,8 @@
 #include <string>
 #include <algorithm>
 #include <functional>
+#include <AVLtree.hpp>
 #include "SumStruct.hpp"
-#include "AVLtree.hpp"
 #include "commands.hpp"
 
 void inputMaps(std::istream& in, novokhatskiy::Tree< int, std::string >& map)
