--- novokhatskiy.danil/S5/stack.hpp
@@ -1,51 +1,0 @@
-#ifndef STACK_HPP
-#define STACK_HPP
-
-#include <stdexcept>
-#include "forward_list.hpp"
-
-namespace novokhatskiy
-{
-  template < class T >
-  class Stack
-  {
-  public:
-    bool operator==(const Stack< T >& other) const
-    {
-      return data_ == other.data_;
-    }
-    bool operator!=(const Stack< T >& other) const
-    {
-      return data_ != other.data_;
-    }
-    T& top()
-    {
-      if (empty())
-      {
-        throw std::invalid_argument("Stack is empty");
-      }
-      return data_.front();
-    }
-    bool empty() const noexcept
-    {
-      return data_.empty();
-    }
-    size_t size() const noexcept
-    {
-      return data_.size();
-    }
-    void push(const T& value)
-    {
-      data_.push_front(value);
-    }
-    void pop()
-    {
-      data_.pop_front();
-    }
-
-  private:
-    ForwardList< T > data_;
-  };
-}
-
-#endif
