+++ novokhatskiy.danil/common/AVLtree.hpp
@@ -4,32 +4,34 @@
 #include <functional>
 #include <cstddef>
 #include <iostream>
+#include <cassert>
+#include <type_traits>
 #include <stdexcept>
 #include <utility>
-#include "AVLtreeNode.hpp"
-#include "queue.hpp"
-#include "stack.hpp"
-#include "constIteratorTree.hpp"
-#include "IteratorTree.hpp"
+#include <stack.hpp>
+#include <queue.hpp>
+#include <AVLtreeNode.hpp>
+#include <constIteratorTree.hpp>
+#include <IteratorTree.hpp>
 
 namespace novokhatskiy
 {
-  template <class Key, class Value, class Compare = std::less<Key>>
+  template < class Key, class Value, class Compare = std::less< Key > >
   class Tree
   {
   public:
-    using node_t = detail::NodeTree<Key, Value>;
-    using v_type = std::pair<Key, Value>;
+    using node_t = detail::NodeTree< Key, Value >;
+    using v_type = std::pair< Key, Value >;
     using constIter = ConstIteratorTree< Key, Value, Compare >;
     using iter = IteratorTree< Key, Value, Compare >;
 
-    Tree() :
+    Tree():
       root_(nullptr),
       size_(0),
       cmp_()
     {}
 
-    Tree(const Tree& other) :
+    Tree(const Tree &other):
       root_(nullptr),
       size_(0),
       cmp_(other.cmp_)
@@ -41,23 +43,24 @@ namespace novokhatskiy
           insert(*begin);
         }
       }
-      catch (const std::exception&)
+      catch (const std::exception &)
       {
         clear();
         throw;
       }
     }
 
-    Tree(Tree&& other) :
+    Tree(Tree &&other) noexcept:
       root_(other.root_),
       size_(other.size_),
       cmp_(std::move(other.cmp_))
     {
+      static_assert(std::is_nothrow_copy_constructible< Compare >::value, "static_assert");
       other.root_ = nullptr;
       other.size_ = 0;
     }
 
-    Tree& operator=(const Tree& other)
+    Tree &operator=(const Tree &other)
     {
       Tree map(other);
       clear();
@@ -65,7 +68,7 @@ namespace novokhatskiy
       return *this;
     }
 
-    Tree& operator=(Tree&& other)
+    Tree &operator=(Tree &&other) noexcept
     {
       clear();
       this->swap(other);
@@ -85,13 +88,13 @@ namespace novokhatskiy
       }
       return f;
     }
-
+    
     template< class F >
     F traverse_lnr(F f) const
     {
       return static_cast< const F >(traverse_lnr(f));
     }
-
+    
     template< class F >
     F traverse_rnl(F f)
     {
@@ -105,13 +108,13 @@ namespace novokhatskiy
       }
       return f;
     }
-
+    
     template< class F >
     F traverse_rnl(F f) const
     {
       return static_cast< const F >(traverse_rnl(f));
     }
-
+    
     template< class F >
     F traverse_breadth(F f)
     {
@@ -137,124 +140,79 @@ namespace novokhatskiy
       }
       return f;
     }
-
+    
     template< class F >
     F traverse_breadth(F f) const
     {
       return static_cast< const F >(traverse_breadth(f));
     }
 
-    node_t* find(node_t* node, const Key& key)
+    iter find(const Key &key)
     {
-      node_t* root = (node->height < 0) ? node->left : node;
-      while (root)
+      node_t *curr = root_;
+      while (curr)
       {
-        if (key == root->value.first)
+        if (cmp_(curr->value.first, key))
         {
-          return root;
+          curr = curr->right;
         }
-        if (cmp_(key, root->value.first))
+        else if (cmp_(key, curr->value.first))
         {
-          if (!root->left)
-          {
-            return root;
-          }
-          root = root->left;
+          curr = curr->left;
         }
         else
         {
-          if (!root->right)
-          {
-            return root;
-          }
-          root = root->right;
-        }
-      }
-      return nullptr;
-    }
-
-    iter find(const Key& key)
-    {
-      node_t* curr = root_;
-      while (curr)
-      {
-        if (curr->value.first == key)
-        {
           return iter(curr);
         }
-        curr = (cmp_(curr->value.first, key)) ? curr->right : curr->left;
       }
       return iter(curr);
     }
 
-    constIter find(const Key& key) const
+    constIter find(const Key &key) const
     {
-      node_t* curr = root_;
+      node_t *curr = root_;
       while (curr)
       {
-        if (curr->value.first == key)
+        if (cmp_(curr->value.first, key))
+        {
+          curr = curr->right;
+        }
+        else if (cmp_(key, curr->value.first))
+        {
+          curr = curr->left;
+        }
+        else
         {
           return constIter(curr);
         }
-        curr = (cmp_(curr->value.first, key)) ? curr->right : curr->left;
       }
       return constIter(curr);
     }
 
     Value& at(const Key& key)
     {
-      node_t* curr = search(key);
-      if (curr)
+      iter curr = find(key);
+      if (curr.node_)
       {
-        return curr->value.second;
+        return curr.node_->value.second;
       }
       throw std::out_of_range("No such element");
     }
 
-    const Value& at(const Key& key) const
+    const Value &at(const Key &key) const
     {
-      node_t* curr = search(key);
-      if (curr)
+      constIter curr = find(key);
+      if (curr.node_)
       {
-        return curr->value.second;
+        return curr.node_->value.second;
       }
       throw std::out_of_range("No such element");
     }
 
-    Value& operator[](const Key& key)
+    Value &operator[](const Key &key)
     {
-      node_t* tmp = search(key);
-      if (tmp)
-      {
-        return tmp->value.second;
-      }
-      else
-      {
-        insert(std::make_pair(key, Value()));
-        node_t* res = search(key);
-        return res->value.second;
-      }
-    }
-
-    const Value& operator[](const Key& key) const
-    {
-      node_t* tmp = search(key);
-      if (tmp)
-      {
-        return tmp->value.second;
-      }
-      else
-      {
-        insert(std::make_pair(key, Value()));
-        node_t* res = search(key);
-        return res->value.second;
-      }
-    }
-
-    node_t* updateHeight(node_t* node)
-    {
-      node->height = 1 + std::max(node->getHeight(node->left), node->getHeight(node->right));
-      return node;
+      auto tmp = std::make_pair(key, Value());
+      return (insert(tmp).first)->second;
     }
 
     size_t size() const noexcept
@@ -273,8 +231,9 @@ namespace novokhatskiy
       root_ = nullptr;
     }
 
-    void swap(Tree& other)
+    void swap(Tree &other) noexcept
     {
+      static_assert(std::is_nothrow_copy_constructible<Compare>::value, "static_assert");
       std::swap(root_, other.root_);
       std::swap(size_, other.size_);
       std::swap(cmp_, other.cmp_);
@@ -284,15 +243,15 @@ namespace novokhatskiy
     {
       iter res = pos;
       res++;
-      node_t* node = nullptr;
-      node_t* delElement = nullptr;
+      node_t *node = nullptr;
+      node_t *delElement = nullptr;
       if (!pos.node_->left || !pos.node_->right)
       {
         delElement = pos.node_;
         node = delElement->parent;
         if (!pos.node_->left && !pos.node_->right)
         {
-          if (node->left = delElement)
+          if (node->left == delElement)
           {
             node->left = nullptr;
           }
@@ -351,7 +310,7 @@ namespace novokhatskiy
       return iter(res.node_);
     }
 
-    iter lower_bound(const Key& key)
+    iter lower_bound(const Key &key)
     {
       iter curr = find(key);
       if (cmp_(curr.node_->value.first, key))
@@ -361,7 +320,7 @@ namespace novokhatskiy
       return curr;
     }
 
-    constIter lower_bound(const Key& key) const
+    constIter lower_bound(const Key &key) const
     {
       iter curr = find(key);
       if (cmp_(curr.node_->value.first, key))
@@ -371,7 +330,7 @@ namespace novokhatskiy
       return curr;
     }
 
-    iter upper_bound(const Key& key)
+    iter upper_bound(const Key &key)
     {
       iter res = lower_bound(key);
       if (res.node_->value.first == key && res != end())
@@ -381,7 +340,7 @@ namespace novokhatskiy
       return res;
     }
 
-    constIter upper_bound(const Key& key) const
+    constIter upper_bound(const Key &key) const
     {
       constIter res = lower_bound(key);
       if (res.node_->value.first == key && res != cend())
@@ -391,17 +350,22 @@ namespace novokhatskiy
       return res;
     }
 
-    std::pair< iter, iter > equal_range(const Key& key)
+    constIter cbeginRight() const
+    {
+      return constIter(lastRight(root_));
+    }
+
+    std::pair< iter, iter > equal_range(const Key &key)
     {
       return std::make_pair(lower_bound(key), upper_bound(key));
     }
 
-    std::pair< constIter, constIter > equal_range(const Key& key) const
+    std::pair< constIter, constIter > equal_range(const Key &key) const
     {
       return std::make_pair(lower_bound(key), upper_bound(key));
     }
 
-    size_t count(const Key& key)
+    size_t count(const Key &key)
     {
       return (find(key) != end());
     }
@@ -426,37 +390,17 @@ namespace novokhatskiy
       return iter();
     }
 
-    constIter cbeginRight() const
+    std::pair< iter, bool > insert(const std::pair< Key, Value >& val)
     {
-      return constIter(lastRight(root_));
-    }
-
-    node_t* search(const Key& key)
-    {
-      node_t* result = search_imp(root_, key);
-      return result;
-    }
-
-    void insert(const std::pair< Key, Value >& val)
-    {
-      try
+      node_t* tmp = search(val.first);
+      bool isInsert = false;
+      if ((tmp && tmp->value.first != val.first) || !tmp)
       {
-        if (root_)
-        {
-          insert_imp(val, root_);
-          ++size_;
-        }
-        else
-        {
-          root_ = new node_t(val);
-          ++size_;
-        }
+        insert_imp(val, tmp);
+        ++size_;
+        isInsert = true;
       }
-      catch (const std::exception&)
-      {
-        clear();
-        throw;
-      }
+      return std::pair< iter, bool >(find(val.first), isInsert);
     }
 
     ~Tree()
@@ -465,7 +409,7 @@ namespace novokhatskiy
     }
 
   private:
-    node_t* root_;
+    node_t *root_;
     size_t size_;
     Compare cmp_;
 
@@ -482,8 +426,8 @@ namespace novokhatskiy
       }
       return res;
     }
-
-    void balance(node_t* node)
+    
+    void balance(node_t *node)
     {
       if (node->height < 0)
       {
@@ -503,7 +447,19 @@ namespace novokhatskiy
       }
     }
 
-    void getBalance(node_t* node)
+    node_t *search(const Key &key) const
+    {
+      node_t *result = search_imp(root_, key);
+      return result;
+    }
+
+    node_t *updateHeight(node_t *node)
+    {
+      node->height = 1 + std::max(node->getHeight(node->left), node->getHeight(node->right));
+      return node;
+    }
+
+    void getBalance(node_t *node)
     {
       if ((node->height > 1) || (node->height < -1))
       {
@@ -529,9 +485,9 @@ namespace novokhatskiy
       }
     }
 
-    node_t* rotate_left(node_t* node)
+    node_t *rotate_left(node_t *node)
     {
-      node_t* newRoot = node->right;
+      node_t *newRoot = node->right;
       node->right = newRoot->left;
       if (newRoot->left)
       {
@@ -544,9 +500,9 @@ namespace novokhatskiy
       return newRoot;
     }
 
-    node_t* rotate_right(node_t* node)
+    node_t *rotate_right(node_t *node)
     {
-      node_t* newRoot = node->left;
+      node_t *newRoot = node->left;
       node->left = newRoot->right;
       if (newRoot->right)
       {
@@ -559,43 +515,39 @@ namespace novokhatskiy
       return newRoot;
     }
 
-    void insert_imp(const std::pair< Key, Value >& value, node_t* node)
+    void insert_imp(const std::pair<Key, Value>& value, node_t* node)
     {
-      try
+      if (!node)
       {
-        if (cmp_(value.first, node->value.first))
+        root_ = new node_t(value);
+      }
+      else if (cmp_(value.first, node->value.first))
+      {
+        if (node->left)
         {
-          if (node->left)
-          {
-            insert_imp(value, node->left);
-          }
-          else
-          {
-            node->left = new node_t(value, node);
-            getBalance(node->left);
-          }
+          insert_imp(value, node->left);
         }
         else
         {
-          if (node->right)
-          {
-            insert_imp(value, node->right);
-          }
-          else
-          {
-            node->right = new node_t(value, node);
-            getBalance(node->right);
-          }
+          node->left = new node_t(value, node);
+          getBalance(node->left);
         }
       }
-      catch (const std::exception&)
+      else
       {
-        clear();
-        throw;
+        if (node->right)
+        {
+          insert_imp(value, node->right);
+        }
+        else
+        {
+          node->right = new node_t(value, node);
+          getBalance(node->right);
+        }
       }
     }
 
-    void clear_imp(node_t* root)
+    void clear_imp(node_t *root)
     {
       if (root)
       {
@@ -606,9 +558,38 @@ namespace novokhatskiy
       }
     }
 
-    node_t* minN(node_t* p) const
+    node_t *find(node_t *node, const Key &key)
     {
-      node_t* res = p;
+      node_t *root = (node->height < 0) ? node->left : node;
+      while (root)
+      {
+        if (key == root->value.first)
+        {
+          return root;
+        }
+        if (cmp_(key, root->value.first))
+        {
+          if (!root->left)
+          {
+            return root;
+          }
+          root = root->left;
+        }
+        else
+        {
+          if (!root->right)
+          {
+            return root;
+          }
+          root = root->right;
+        }
+      }
+      return nullptr;
+    }
+
+    node_t *minN(node_t *p) const
+    {
+      node_t *res = p;
       if (!res)
       {
         return nullptr;
@@ -620,7 +601,7 @@ namespace novokhatskiy
       return res;
     }
 
-    node_t* search_imp(node_t* node, const Key& key)
+    node_t* search_imp(node_t* node, const Key& key) const
     {
       if (!node)
       {
@@ -628,11 +609,25 @@ namespace novokhatskiy
       }
       else if (cmp_(key, node->value.first))
       {
-        return search_imp(node->left, key);
+        if (node->left)
+        {
+          return search_imp(node->left, key);
+        }
+        else
+        {
+          return node;
+        }
       }
       else if (cmp_(node->value.first, key))
       {
-        return search_imp(node->right, key);
+        if (node->right)
+        {
+          return search_imp(node->right, key);
+        }
+        else
+        {
+          return node;
+        }
       }
       else
       {
